backend/requirements.txt
#Core AWS SDK for Python (required for all Lambda functions)
boto3

#Optional: pydantic for structured data validation (good for security)
pydantic[email]>=2.7.0
pydantic-settings>=2.12.0

#Logging library for structured logs (helps with audit trails)
pyjwt

#Web framework for building APIs
fastapi

#ASGI adapter for AWS Lambda
mangum

#Testing framework (optional, but useful for development)
pytest

#Library to load environment variables from .env file (for local development)
python-dotenv

#Library for generating secure tokens
python-json-logger
#HTTP client library for async API calls (WhatsApp/Instagram)
httpx>=0.25.0

#HTTP client library for sync API calls (OAuth, external integrations)
requests>=2.31.0

#Password hashing library
bcrypt>=4.0.0

#PDF generation library for transaction summaries
reportlab>=4.0.0

#QR code generation for order tracking
qrcode[pil]>=7.4.0
